Initial commit of AST generation from block
This makes quite a bit of use of Bindings, blocks/methods and the methods on
them. Effectively, we intercept all method calls and try to wrap them in our
AST if there is exactly one parameter slot we can fill in the pipeline.

If this is not the case, we simply pass it on to the original receiver, allowing
for normal method calls, and reasonable failure messages.