Allow procs and lambdas to be used as inputs and destinations in a pipeline
When given as an input, a proc is treated as a lambda, otherwise it is treated as a AST::Block, which is a destination-only node. To be reifiable, a proc must be unary, any other arity results in an UnreifiableNodeError. AST::Block represents both lambdas and procs.

N.B.: proc/lambda will not be captured by AST::MethodCall because they are nil-arity methods and not reifiable, they will be normally resolved in the scope of the block.