Add concept of "object proxies" to syntax
The idea is that there are cases where you want to use a method on another receiver besides the current self in a pipeline definition. e.g.:

```rb
input(25) >> Math.sqrt \
          >> Math.hypot(12) \
          >> 1.upto
```

To get back a range equivalent to `(1..13)`. Unfortunately, it is infeasible to modify constant resolution and intercepting a literal, even if possible, would be awful. This commit introduces the `on` syntax to pipelines. This will wrap the given object in a proxy that will intercept *most* method calls and turn them into AST::MethodCall nodes under the same constraints as any other AST::MethodCall generator. The previous example becomes:

```rb
input(25) >> on(Math).sqrt \
          >> on(Math).hypot(12) \
	  >> on(1).upto
```

Additionally, there is some cleanup regarding usage of BasicObject, providing a custom Parser::BasicObject that forwards some methods to Kernel.